Script generates links/thumbs for themes in README
Includes thumbs an GH links, but not themes.gohugo.io because it's not
up yet. There's a TODO in the code for when it is.

README.md is now generated by this generate-readme script. There is now
a warning at the top of the (rendered) README.md noting that the code
there is auto-generated and will be overwritten.

Therefore, README.tmpl.md provides a template for the static part, which
now contains the contents of the former README.md

README.md now contains the former version of README.md with the
generated theme list appended to it, preceeded by the warning noted
above.

The current list has some outdated links (at least one repository did
add a thumbnail), so the submodules should probably be updated soon.

Fixes #40