Clean Up Repo from Unmaintained Themes (#794)
This commit removes unmaintained themes from the repo that have been for a number of years in the blackList and noDemoList of the Build Script.

These are themes that either do not comply with Hugo Themes requirements like the Persona theme or have been unmaintained for years.

There is no need to keep them in the repo.