Simplify theme submission process (#853)
This PR emphasizes the need of theme submissions to fulfill the criteria mentioned in the README before being opened.

Also this PR will make it easier for us to maintain this repo since from now on authors who are not experienced with Hugo will be directed to the Support Forum for their questions. 

It just is not viable for us to continue tutoring inexperienced theme authors. Especially since most of the time the problems that are present in the submissions already have detailed fixes mentioned in the README.

I am not merging this PR since I have asked for reviews from the other maintainers.

P.S. Also this PR restores the filename to `theme-submission.md` that changes due to an error on my part a while back due to a script.